chore: 폴더명 오타 수정 (uttils->utils)
