+++ __tests__/LottoTest.js
@@ -1,6 +1,6 @@
 import Lotto from '../src/Lotto.js';
 import { ERROR_MESSAGE } from '../src/constant';
-import { getErrorMessage } from '../src/uttils/MessageFactory.js';
+import { getErrorMessage } from '../src/utils/MessageFactory.js';
 
 describe('로또 클래스 테스트', () => {
   test('로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.', () => {
