--- src/uttils/index.js
@@ -1,13 +1,0 @@
-import { getErrorMessage, throwError, printMessage } from './MessageFactory';
-import { getBonusBallNumber, getLottoRandomNumbers } from './RandomNumbers';
-import { validateNumberRange, hasNoReset } from './Validate';
-
-export {
-  getBonusBallNumber,
-  getErrorMessage,
-  getLottoRandomNumbers,
-  hasNoReset,
-  printMessage,
-  throwError,
-  validateNumberRange,
-};
